Split up DB class and allow for login gui to authenticate
